+++ ricaun.Revit.UI/RibbonItemDataExtension.cs
@@ -1,4 +1,5 @@
 ﻿using Autodesk.Revit.UI;
+using ricaun.Revit.UI.Utils;
 using System.Windows.Media;
 
 namespace ricaun.Revit.UI
@@ -150,6 +151,8 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource image) where TRibbonItem : RibbonItemData
         {
+            image = image.GetThemeImageSource(RibbonThemeUtils.IsLight);
+
             if (ribbonItem is ButtonData ribbonButton)
                 ribbonButton.Image = image?.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
 
@@ -174,6 +177,8 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItemData
         {
+            largeImage = largeImage.GetThemeImageSource(RibbonThemeUtils.IsLight);
+
             if (ribbonItem is ButtonData ribbonButton)
             {
                 ribbonButton.LargeImage = largeImage?.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });